move trigger_reminders.py up one level
