+++ stopcovid/drill_progress/trigger_reminders.py
